instalado toastr, agregado conectividad, notificacion, actializado pwanofier

+++ src/app/pages/profile/paymentmethod/paymentmethod.component.ts
@@ -12,6 +12,7 @@ import { ActivatedRoute } from '@angular/router';
 import { LoadingComponent } from '../../../shared/loading/loading.component';
 import Swal from 'sweetalert2';
 import { TranslateModule } from '@ngx-translate/core';
+import { ToastrService } from 'ngx-toastr';
 
 @Component({
   selector: 'app-paymentmethod',
@@ -31,31 +32,32 @@ import { TranslateModule } from '@ngx-translate/core';
 })
 export class PaymentmethodComponent {
   pageTitle = 'Payment Methods';
-  public tiposdepagos: PaymentMethod [] = [];
-  public tiposdepagosuser: PaymentMethod [] = [];
+  public tiposdepagos: PaymentMethod[] = [];
+  public tiposdepagosuser: PaymentMethod[] = [];
   public user!: any;
-  isLoading:boolean = false;
-  user_id!:number;
-  tipoSeleccionado:any;
-  pagoSeleccionado:any;
+  isLoading: boolean = false;
+  user_id!: number;
+  tipoSeleccionado: any;
+  pagoSeleccionado: any;
 
   public paymentMForm!: FormGroup;
 
-  username!:PaymentMethod;
-  bankAccountType!:string;
-  bankName!:string;
-  bankAccount!:string;
-  ciorif!:string;
-  phone!:string;
-  email!:string;
-  tipo!:string;
-  id!:number;
+  username!: PaymentMethod;
+  bankAccountType!: string;
+  bankName!: string;
+  bankAccount!: string;
+  ciorif!: string;
+  phone!: string;
+  email!: string;
+  tipo!: string;
+  id!: number;
 
   constructor(
     private authService: AuthService,
     private paymentMService: PaymentmethodService,
     private fb: FormBuilder,
     private activatedRoute: ActivatedRoute,
+    private toastr: ToastrService,
   ) {
     this.user = this.authService.getLocalStorage();
   }
@@ -82,30 +84,24 @@ export class PaymentmethodComponent {
   }
 
 
-  selectedTypeEdit(tipo:any){
+  selectedTypeEdit(tipo: any) {
     this.pagoSeleccionado = tipo.value;
     // console.log(this.pagoSeleccionado);
-}
+  }
 
-selectedType(tipodepago:any){
+  selectedType(tipodepago: any) {
     this.tipoSeleccionado = tipodepago;
     // console.log(this.tipoSeleccionado);
-}
+  }
 
 
-cambiarStatus(tipodepago:any){
+  cambiarStatus(tipodepago: any) {
     let VALUE = tipodepago.status;
     // console.log(VALUE);
-    
+
     this.paymentMService.updateStatus(tipodepago.status, tipodepago._id).subscribe(
-      (resp:any) =>{
-        // console.log(resp);
-        Swal.fire('Actualizado', `actualizado correctamente`, 'success');
-        // this.toaster.open({
-        //   text:'Producto Actualizado!',
-        //   caption:'Mensaje de Validación',
-        //   type:'success',
-        // })
+      (resp: any) => {
+        this.toastr.success('Actualizado', `actualizado correctamente`)
         this.getPaymentMethodByUserId();
       }
     )
@@ -113,20 +109,20 @@ cambiarStatus(tipodepago:any){
 
 
 
-save(){
+  save() {
 
     let data = {
       tipo: this.tipo,
       bankAccountType: this.bankAccountType,
       bankName: this.bankName,
       bankAccount: this.bankAccount,
-      ciorif:this.ciorif,
-      phone:this.phone,
+      ciorif: this.ciorif,
+      phone: this.phone,
       email: this.email,
       user: this.user.uid
     }
-    this.paymentMService.createPaymentmethod(data).subscribe((resp:any)=>{
-      // console.log(resp);
+    this.paymentMService.createPaymentmethod(data).subscribe((resp: any) => {
+      this.toastr.success('Exito', `Creado correctamente`)
       this.tipo = '';
       this.bankAccountType = '';
       this.bankName = '';
@@ -138,14 +134,31 @@ save(){
     })
   }
 
-deleteTipoPago(tiposdepago:any){
 
-    this.paymentMService.deletePaymentmethod(tiposdepago.id).subscribe(
-      (resp:any) =>{
-        this.getPaymentMethodByUserId();
-        
-      });
-    
+  deleteTipoPago(tiposdepago: any) {
+    Swal.fire({
+      title: 'Estas Seguro?',
+      text: "No podras recuperarlo!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Si, Borrar!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.paymentMService.deletePaymentmethod(tiposdepago.id).subscribe(
+          response => {
+            this.getPaymentMethodByUserId();
+          }
+        );
+        Swal.fire(
+          'Borrado!',
+          'El Archivo fue borrado.',
+          'success'
+        )
+        this.ngOnInit();
+      }
+    });
   }
 
 }
